Polishing

+++ src/main/java/org/oagi/srt/gateway/http/api/cc_management/repository/ReleaseManifestRepository.java
@@ -67,31 +67,31 @@ public class ReleaseManifestRepository {
                 .fetchOne();
     }
 
-    public List<AsccReleaseManifestRecord> getAsccReleaseManifestByToAsccpId(long toAsccpId, long releaseId) {
+    public List<AsccReleaseManifestRecord> getAsccReleaseManifestByToAsccpId(ULong toAsccpId, ULong releaseId) {
         return dslContext.selectFrom(ASCC_RELEASE_MANIFEST)
-                .where(and(ASCC_RELEASE_MANIFEST.TO_ASCCP_ID.eq(ULong.valueOf(toAsccpId)),
-                        ASCC_RELEASE_MANIFEST.RELEASE_ID.eq(ULong.valueOf(releaseId))))
+                .where(and(ASCC_RELEASE_MANIFEST.TO_ASCCP_ID.eq(toAsccpId),
+                        ASCC_RELEASE_MANIFEST.RELEASE_ID.eq(releaseId)))
                 .fetch();
     }
 
-    public List<BccReleaseManifestRecord> getBccReleaseManifestByToBccpId(long toBccpId, long releaseId) {
+    public List<BccReleaseManifestRecord> getBccReleaseManifestByToBccpId(ULong toBccpId, ULong releaseId) {
         return dslContext.selectFrom(BCC_RELEASE_MANIFEST)
-                .where(and(BCC_RELEASE_MANIFEST.TO_BCCP_ID.eq(ULong.valueOf(toBccpId)),
-                        BCC_RELEASE_MANIFEST.RELEASE_ID.eq(ULong.valueOf(releaseId))))
+                .where(and(BCC_RELEASE_MANIFEST.TO_BCCP_ID.eq(toBccpId),
+                        BCC_RELEASE_MANIFEST.RELEASE_ID.eq(releaseId)))
                 .fetch();
     }
 
-    public List<AsccReleaseManifestRecord> getAsccReleaseManifestByFromAccId(long fromAccId, long releaseId) {
+    public List<AsccReleaseManifestRecord> getAsccReleaseManifestByFromAccId(ULong fromAccId, ULong releaseId) {
         return dslContext.selectFrom(ASCC_RELEASE_MANIFEST)
-                .where(and(ASCC_RELEASE_MANIFEST.FROM_ACC_ID.eq(ULong.valueOf(fromAccId)),
-                        ASCC_RELEASE_MANIFEST.RELEASE_ID.eq(ULong.valueOf(releaseId))))
+                .where(and(ASCC_RELEASE_MANIFEST.FROM_ACC_ID.eq(fromAccId),
+                        ASCC_RELEASE_MANIFEST.RELEASE_ID.eq(releaseId)))
                 .fetch();
     }
 
-    public List<BccReleaseManifestRecord> getBccReleaseManifestByFromAccId(long fromAccId, long releaseId) {
+    public List<BccReleaseManifestRecord> getBccReleaseManifestByFromAccId(ULong fromAccId, ULong releaseId) {
         return dslContext.selectFrom(BCC_RELEASE_MANIFEST)
-                .where(and(BCC_RELEASE_MANIFEST.FROM_ACC_ID.eq(ULong.valueOf(fromAccId)),
-                        BCC_RELEASE_MANIFEST.RELEASE_ID.eq(ULong.valueOf(releaseId))))
+                .where(and(BCC_RELEASE_MANIFEST.FROM_ACC_ID.eq(fromAccId),
+                        BCC_RELEASE_MANIFEST.RELEASE_ID.eq(releaseId)))
                 .fetch();
     }
